+++ examples/src/CLOCK_16x02/main.cpp
@@ -23,7 +23,7 @@
 
 // Section: Globals
 // myLCD(rows , cols , PCF8574 I2C address, I2C speed)
-HD44780PCF8574LCD myLCD( 2, 16, 0x27, 0); // instantiate an object
+HD44780PCF8574LCD myLCD( 2, 16, 0x27, BCM2835_I2C_CLOCK_DIVIDER_626); // instantiate an object
 
 // Section: Function Prototypes
 bool setup(void);
@@ -67,6 +67,7 @@ bool setup(void) {
 	if (!myLCD.LCD_I2C_ON())
 	{
 		std::cout << "Error 1202: bcm2835_i2c_begin :Cannot start I2C, Running as root?" << std::endl;
+		bcm2835_close(); // Close the library
 		return false;
 	}