+++ examples/src/TEST_I2C_16x02/main.cpp
@@ -17,18 +17,18 @@
 
 // Section: Globals
 // myLCD(rows , cols , PCF8574 I2C address, I2C speed)
-HD44780PCF8574LCD myLCD(2, 16, 0x27, 0); // instantiate an object
+HD44780PCF8574LCD myLCD(2, 16, 0x27, BCM2835_I2C_CLOCK_DIVIDER_626 ); // instantiate an object
 
 // Section: Function Prototypes
 bool setup(void);
-void helloWorld(void);
+void test(void);
 void endTest(void);
 
 // Section: Main Loop
 int main(int argc, char **argv)
 {
 	if (!setup()) return -1;
-	helloWorld();
+	test();
 	endTest();
 	return 0;
 } 
@@ -49,15 +49,17 @@ bool setup(void) {
 	if (!myLCD.LCD_I2C_ON())
 	{
 		std::cout << "Error 1202: bcm2835_i2c_begin :Cannot start I2C, Running as root?" << std::endl;
+		bcm2835_close(); // Close the library
 		return false;
 	}
 	
 	bcm2835_delay(1000);
-	myLCD.LCDDebugSet(true); // Turn debug messages on 
+	myLCD.LCDDebugSet(false); // Turn debug messages on 
 	myLCD.LCD_I2C_SetSpeed();
 	if (myLCD.LCDCheckConnection() != 0)
 	{
 		std::cout << "Error 1203: bcm2835_i2c_begin :LCD not on bus?" << std::endl;
+		bcm2835_close(); // Close the library
 		return false;
 	}
 	
@@ -66,20 +68,22 @@ bool setup(void) {
 	std::cout << "HD44780_LCD_RPI lib Version Num :"  << myLCD.LCDVerNumGet() << std::endl;
 	std::cout << "Debug status is : " << (myLCD.LCDDebugGet() ? "On" : "Off") << std::endl ;
 	std::cout <<  "Backlight status is : " << (myLCD.LCDBackLightGet() ? "On" : "Off") << std::endl ;
-	std::cout << "I2C Debug Error : " << myLCD.LCDI2CErrorGet() << std::endl; // Print I2C error flag
-	std::cout << "I2C Error Timeout mS : " << myLCD.LCDI2CErrorTimeoutGet() << std::endl; // Print I2C Timeout
-	
+	std::cout << "I2C Debug Error : " << +myLCD.LCDI2CErrorGet() << std::endl; // Print I2C error flag
+	std::cout << "I2C Error Timeout mS : " << myLCD.LCDI2CErrorTimeoutGet() << std::endl; // Print I2C error Timeout
+	std::cout << "I2C Error retry attempts counts: " << +myLCD.LCDI2CErrorRetryNumGet() << std::endl; // Print I2C error retry count
 	myLCD.LCDInit(myLCD.LCDCursorTypeOn);
 	myLCD.LCDClearScreen();
 	return true;
 }
 
-void helloWorld(void) 
+void test(void) 
 {
+	std::cout << "Test" << std::endl;
 	char testString[] = "I2C Test";
 	myLCD.LCDGOTO(myLCD.LCDLineNumberOne, 0);
 	myLCD.LCDSendString(testString);
 	bcm2835_delay(5000);
+	myLCD.LCDClearScreen();
 }